Bugfix: 모임 관리 권한 수정 search 에러 수정
