+++ src/utils/wait.ts
@@ -1,0 +1,3 @@
+export const wait = (timeout: any) => {
+  return new Promise(resolve => setTimeout(resolve, timeout));
+};
